2 new script Actions
82,nn => Decrease Current Trigger Weight
83,nn => Increase Current Trigger Weight

Unlike script action 49,0 that only is evaluated at the end of the script execution these 2 updates the current weight of the trigger that belongs the Team in the moment that is executed the script action.
In both scripts are positive integers. if the nn value is 0 then AITriggerSuccessWeightDelta & AITriggerFailureWeightDelta are used for updating this value.