New actions
103,nn >>> New modifier. Set AI safe distance for "Moving to Techno" script actions
The action finish when the distance of security against the selected target is reached. Only works for the first execution of the "Moving To Techno" actions after this action.
If not set, the "Moving To Techno" actions use the CloseEnough value in rulesmd.ini.

New Attack mission actions 104 & 105: Given a list from [AITargetType] this action will select randomly 1 target and try to find that object in the map. If not found the action will jump to the script line.

Now the [AITargetType] values parsing is more permissive, not crashing if the ID are wrong.

Now the attack actions are more compatible with Spies.