Script Action 93 backported from PR 296
When executed before a new Attack ScriptType Actions like `74-81` and `84-91` the TeamType will remember that must be rewarded increasing the current Weight of the AI Trigger when the TeamType Target was killed by any of the Team members. The current Weight will never surprass the Minimum Weight and Maximum Weight limits of the AI Trigger. The second parameter is a positive value. 

Only affects the first (new) attack script action ran after this action.

In `aimd.ini`:

[SOMESCRIPTTYPE]  ; ScriptType
x=93,n            ; integer n=0