Script Action 94: Pick Random Script
Script Action 94 Backported from PR 296.

When executed it picks a random Script Type and replaces the current script by the new picked.

The second parameter is a 0-based index for the [AIScriptsList] section that specifies the list of possible [ScriptTypes] that can be evaluated. The new [AIScriptsList] section must be declared in RulesMD.ini for making this script work:

[AIScriptsList] ; List of ScriptTypes
0=ScriptType N,ScriptType N+1, ScriptType N+1,... ScriptType N+M
1= ...
2= ...
...
small code style change


small doc fix