- Added support for Iron Courtain transfer.
- Added support for occupiers/passengers transfer. 
This means Occupiers in the case of structures and Passengers for vehicles. If during the transfer something is invalid the unit will be ejected instead of inserted as passenger/occupier.

- Fixed some deploy cases.

- Convert.EjectPassengers=false
By default passengers/occupiers will be transfered into the target. If set to "true" the passengers/occupiers will be ejected in the deploy process.