[Minor] Attempt to fix the ReloadInTransport crash (#1319)
Techno with ReloadInTransport=yes suffered from a crash when the
transporter's owner is changed (i.e. being mind controlled). This is
because it'll move the techno from the original owner's
OwnedTransportReloader list to the new owner's, which is actually wrong
when the transporter doesn't have Passengers.SyncOwner=yes. Attempted to
fix this issue and unified the conditions of ReloadInTransport

Co-authored-by: chaserli <914137150@qq.com>