Voxel Light Source Position Customization (#1336)
TODO
- [x] transform into world coords
- [x] docs

Allows customizing the direction of voxel light in order to provide
better visuals ingame.

In `rulesmd.ini`
```ini
[AudioVisual]
VoxelLightSource=0.02,-0.69,0.36
VoxelShadowLightSource=0.02,-0.69,0.36
```

The values are unset by default. The example values are according to the
convo in Discord where @thomassneddon and CCrisp mentioned that those
coords correspond are used to produce the following (apparently in
radians? not sure)

> This is the value of projected light source direction i mentioned
above, and as a result, the world space direction of YR VXL light is
approx 0.20, -0.91, -0.36

On stock voxels, VPL, palette this looks like this, vanilla vs the
above:


![vanilla](https://github.com/user-attachments/assets/074058b8-6a7c-4eda-af03-1877b871e64f)

![tweaked](https://github.com/user-attachments/assets/24c4d434-3a7d-4c24-9308-2447d4a05cdf)

---------

Co-authored-by: MortonPL <bartm12@wp.pl>
Co-authored-by: Starkku <Starkku@users.noreply.github.com>