More aircraft actions enhancement for #1366 (#1484)
1. Fix aircrafts planning waypoint problem - At the end of each task,
all units with planning waypoints will generate an mega event to the
next planning waypoint and respond after 1 frame. This is really no
problem for most units, but for aircrafts, they will always return to
the airport after each task. So they will try to return to the airport
in the frame before it respond to event. There will be looked like a
"shaking", if the `Speed` and `ROT` of the aircraft are fast. To solve
this problem, I check the waypoints of aircrafts and let them wait
before returning.
2. Fix a vanilla issue - Now aircrafts can fly towards the water surface
normally.

---------

Co-authored-by: Kerbiter <crabiter@vivaldi.net>