Transfer Chinese translation&change QQ group ID to plain text
The Chinese document homepage has been improved, so there is no longer a need to use English/Chinese as a proxy for navigation in English documents. The group ID is separated into a plain text after the link text to avoid various network issues, as it can be directly copied without accidental contact.