[Minor] Shield armor inheritance customization (#1403)
Further customize shield's `InheritArmorFromTechno` logic:
- `InheritArmor.Allowed` lists TechnoTypes whose armor can be
overridden. If empty, any TechnoType not listed in
`InheritArmor.Disallowed` is okay.
- `InheritArmor.Disallowed` lists TechnoTypes whose armor can't be
overridden. If empty, any TechnoTypes are okay as long as
`InheritArmor.Allowed` is empty or they are listed on it.

In `rulesmd.ini`:
```ini
[SOMESHIELDTYPE]                            ; ShieldType name
InheritArmor.Allowed=                       ; list of TechnoTypes
InheritArmor.Disallowed=                    ; list of TechnoTypes
```