[Customized] Projectile subject to ground check before firing (#1291)
This is a feature split from #1261 .
It may have similarities with other parts.

If set `SubjectToGround` to true ,it will predict will predict the
height of the connecting straight line from the bullet's source
coordinates to target coordinates. If the predicted height is lower than
the ground height of the current predicted position, the firer will also
consider targets blocked by such obstacles as out of range and will
attempt to reposition themselves.

In `rulesmd.ini`:
```ini
[SOMEPROJECTILE]              ; Projectile
SubjectToGround=false         ; boolean
```

---------

Co-authored-by: Starkku <starkku.ppmod@gmail.com>