[Simple Fix] Several repairs related to passengers boarding vehicles (#1613)
Phobos fix:
Fixed an issue that passengers would not gather at the board position
when the transport ship came to pick them up when `NoQueueUpToEnter` was
enabled.
Fixed an issue that overlap may still occur when `NoQueueUpToEnter` was
enabled.
Fixed an issue that the sound will continue to play if passengers are
unable to unload all at once when `NoQueueUpToEnter` was enabled.
Fixed an issue that foots will leave impassable invisible barrier when
boarding vehicles on the bridge, which can later cause the game to
freeze when `NoQueueUpToEnter` was enabled.
Fixed an issue that foots under the bridge are allowed to board vehicles
on the bridge when `NoQueueUpToEnter` was enabled.

Vanilla fix:
Fixed an issue that the first passenger who call the transport ship no
longer board the transport ship when the land units call for boarding.
Fixed an issue that impassable invisible barrier generated by the
behavior of infantry continuously entering vehicles.
Fixed an issue that teleport units board transport vehicles on the
bridge will create an impassable invisible barrier, which may cause the
game to freeze or even crash

Splits from #1602 .