Relicense under GPL v3 (#1567)
In order to accept contributions containing code licensed under GPL, it
is required to relicense Phobos under GPL, as well.
This does not present any issues, as the extra permissions provided by
LGPL are not required for Phobos.
This PR relicenses Phobos under GPL v3.

Adjustments from the Chinese community are required to update the
Chinese documentation to reference GPL v3.

---------

Co-authored-by: 九千天华 <1065703286@qq.com>