[Minor] Fix `KeepTargetOnMove` will stuck unit status (#1538)
The problem arises in:

When a unit with `KeepTargetOnMove` moves after attacking a target, it
will enter `Mission::Guard` instead of `Mission::Attack` when it stops,
which will result in it not only not pursuing but also not resetting its
target when the target moves away from it.

I have added an option to clear the target or not. If not, the unit will
be back into `Mission::Attack` so that it can start pursuing its target.
- `KeepTargetOnMove.NoMorePursuit` controls whether the unit will
restart chasing the target for attack when it stops again, otherwise it
will clear the target when it moves away.