[Minor] Customizing bridge fall damageFalling down damage (#1421)
![坠落3](https://github.com/user-attachments/assets/9cfed9d4-d19f-423b-b568-17026bbe4f10)

![坠落4](https://github.com/user-attachments/assets/6a128485-5ab0-42b1-88e0-091b92847f94)

Now you can customize the damage a unit receives when it falls from a
bridge.
- `FallingDownDamage` customizes the damage a unit receives at the end
of a fall. It can be a percentage or an integer.
- `FallingDownDamage.Water` customizes the damage a unit receives when
it falls onto the water. Defaults to `FallingDownDamage`.
- If it is a negative percentage, corresponding damage will be dealt
based on the current health of the unit.

In `rulesmd.ini`:
```ini
[SOMETECHNO]                    ; TechnoType
FallingDownDamage=              ; integer / percentage
FallingDownDamage.Water=        ; integer / percentage
```
---------
现在可以自定义单位从桥上坠落下来时受到的伤害。
- `FallingDownDamage`自定义单位坠落后受到的伤害。它可以是百分比，也可以是正数。
- `FallingDownDamage.Water`自定义自定义单位坠落到水面时受到的伤害。默认为`FallingDownDamage`。
- 如果是负百分比的话，它会根据单位的当前血量赋予对应百分比的伤害。