More code
Improved & splitted the code in auxiliar functions.
Now the SW can modify the house points.
Now a new General tag is added for default value when the tag isn't established in the techno.
Added hook SuperClass_IsReady_BattlePoints for disabling the use of the SW if it doesn't have enough points (in reality is a rewrite of that IsReady() method adding this check) .