[Minor] ExpireWeapon/RevengeWeapon/ReflectDamage UseInvokerAsOwner (#1725)
- `ExpireWeapon.UseInvokerAsOwner` can be used to set the house and
TechnoType that created the effect (e.g firer of the weapon that applied
it) as the weapon's owner & invoker instead of the object the effect is
attached to.
- `RevengeWeapon.UseInvokerAsOwner` can be used to set the house and
TechnoType that created the effect (e.g firer of the weapon that applied
it) as the weapon's owner & invoker instead of the object the effect is
attached to.
- `ReflectDamage.UseInvokerAsOwner` can be used to set the house and
TechnoType that created the effect (e.g firer of the weapon that applied
it) as the reflected damage's owner & invoker instead of the object the
effect is attached to.

In `rulesmd.ini`:
```ini
[SOMEATTACHEFFECT]                                 ; AttachEffectType
ExpireWeapon.UseInvokerAsOwner=false               ; boolean
RevengeWeapon.UseInvokerAsOwner=false              ; boolean
ReflectDamage.UseInvokerAsOwner=false              ; boolean
```