[Minor] Damage.DealtByInvoker tweak & Damage.ApplyFIrepowerMult (#1728)
previous `Damage.DealtByInvoker` doesn't really apply to anim warhead,
this PR makes it inline with anim weapon. Notice that this might change
the behavior of anim weapon and warhead a bit, but I think this should
be the correct way. Any other feedbacks and reviews are welcome

`Damage.ApplyFirepowerMult` determines whether or not firepower
modifiers from the animation's invoker are applied on the damage dealt
from this animation, if exists