[Minor] Fix create building map trigger action (#1848)
Fixes an issue with map trigger action 125 causing buildups to play
incorrectly sometimes. A separate issue remains where any building that
is mid-buildup will interrupt buildup if it changes owner but this is
not exclusive to buildings created through the trigger action.

Also re-enabled use fo the third parameter but this time it turns
buildups off even if otherwise available if set to != 0.