[Customized] Allow the aircraft to enter area guard mission and not crash immediately without any airport (#1844)
- When a `guard` command (`[G]` by default) or a `area guard` command
(`[Ctrl]+[Alt]`) is issued, the aircraft will search for targets around
the position (for `guard` is the current location, for `area guard` is
the target position) and return immediately when ammos are depleted.
- If the target is not found, or if there is still ammo when the target
is destroyed, it will continue to hover over the guarded area.
- `ExtendedAircraftMissions.FastScramble` controls whether the aircraft
can scramble when its airport has been destroyed.
- `ExtendedAircraftMissions.UnlandDamage` controls the damage suffered
by the aircraft every 4 frames when there is no airport for the aircraft
to land. If the value is negative, it will crash immediately. Not
recommended to use when `ExtendedAircraftMissions` is not enabled.


In `rulesmd.ini`:
```ini
[General]
ExtendedAircraftMissions.UnlandDamage=-1  ; integer

[SOMEAIRCRAFT]                            ; AircraftType
ExtendedAircraftMissions.FastScramble=    ; boolean, default to [General] -> ExtendedAircraftMissions
ExtendedAircraftMissions.UnlandDamage=    ; integer, default to [General] -> ExtendedAircraftMissions.UnlandDamage
```


![飞机守卫](https://github.com/user-attachments/assets/7ed4c69c-e5e7-41f6-97d6-8f45d910adb3)