minor optimization on Nullable's `Get()` function
When using `Get(some codes for default value)` to acquire a Nullable, it'll execute the codes for default value even if this Nullable is set, becoming an extra cost in such case.

While it's just reading variables or simple calculation in most cases, some of theses calculation are a bit more complex. This commit make such cases use `isset()` to check if default values are needed first, which cost much less. For cases where the Nullable is not existed, it won't add extra cost technically since `Get()` also has an `isset()` within, which is now omited.