Fix owner change messing up building buildups if occuring simultaneously (#1943)
Fixes a vanilla bug that causes owner change on buildings during buildup
to instantly clear the buildup state. It was only working partially for
buildings created by trigger action 125 (BState flag was not set to idle
which caused buildup anims to loop indefinitely even though mission was
reset), this fix should fix any remaining problems with that for mappers
who use it on conjunction with house changing as well.