Map event 601 behaviour change (#1988)
By default the Map Event 601 returns TRUE as soon as you click in the
sidebar for producing the techno instead of return TRUE when is placed.

The expected behaviour was like other map events that checks if exists
in the map.

Since is a change/fix shouldn't be necessary more documentation about
this.